Extract maps code to module
