+++ collect_app/src/main/java/org/odk/collect/android/application/Collect.java
@@ -50,7 +50,7 @@ import org.odk.collect.geo.GeoDependencyComponent;
 import org.odk.collect.geo.GeoDependencyComponentProvider;
 import org.odk.collect.geo.GeoDependencyModule;
 import org.odk.collect.geo.ReferenceLayerSettingsNavigator;
-import org.odk.collect.geo.maps.MapFragmentFactory;
+import org.odk.collect.maps.MapFragmentFactory;
 import org.odk.collect.location.GpsStatusSatelliteInfoClient;
 import org.odk.collect.location.LocationClient;
 import org.odk.collect.location.satellites.SatelliteInfoClient;
