+++ collect_app/src/main/java/org/odk/collect/android/geo/MapboxMapFragment.java
@@ -64,8 +64,8 @@ import org.odk.collect.android.geo.MbtilesFile.MbtilesException;
 import org.odk.collect.android.injection.DaggerUtils;
 import org.odk.collect.android.location.client.MapboxLocationCallback;
 import org.odk.collect.android.utilities.MapFragmentReferenceLayerUtils;
-import org.odk.collect.geo.maps.MapFragment;
-import org.odk.collect.geo.maps.MapPoint;
+import org.odk.collect.maps.MapFragment;
+import org.odk.collect.maps.MapPoint;
 
 import java.io.File;
 import java.io.IOException;
