+++ collect_app/src/main/java/org/odk/collect/android/geo/OsmDroidMapConfigurator.java
@@ -11,7 +11,7 @@ import com.google.common.collect.ImmutableSet;
 
 import org.odk.collect.android.R;
 import org.odk.collect.android.preferences.PrefUtils;
-import org.odk.collect.geo.maps.MapFragment;
+import org.odk.collect.maps.MapFragment;
 import org.odk.collect.shared.settings.Settings;
 
 import java.io.File;
