+++ geo/src/main/java/org/odk/collect/geo/DaggerSetup.kt
@@ -14,11 +14,11 @@ import org.odk.collect.geo.geopoint.GeoPointMapActivity
 import org.odk.collect.geo.geopoint.GeoPointViewModelFactory
 import org.odk.collect.geo.geopoint.LocationTrackerGeoPointViewModel
 import org.odk.collect.geo.geopoly.GeoPolyActivity
-import org.odk.collect.geo.maps.MapFragmentFactory
 import org.odk.collect.geo.selection.SelectionMapFragment
 import org.odk.collect.location.LocationClient
 import org.odk.collect.location.satellites.SatelliteInfoClient
 import org.odk.collect.location.tracker.LocationTracker
+import org.odk.collect.maps.MapFragmentFactory
 import org.odk.collect.permissions.PermissionsChecker
 import javax.inject.Singleton
