+++ geo/src/main/java/org/odk/collect/geo/selection/SelectionMapFragment.kt
@@ -20,10 +20,10 @@ import org.odk.collect.geo.GeoDependencyComponentProvider
 import org.odk.collect.geo.R
 import org.odk.collect.geo.ReferenceLayerSettingsNavigator
 import org.odk.collect.geo.databinding.SelectionMapLayoutBinding
-import org.odk.collect.geo.maps.MapFragment
-import org.odk.collect.geo.maps.MapFragment.ReadyListener
-import org.odk.collect.geo.maps.MapFragmentFactory
-import org.odk.collect.geo.maps.MapPoint
+import org.odk.collect.maps.MapFragment
+import org.odk.collect.maps.MapFragment.ReadyListener
+import org.odk.collect.maps.MapFragmentFactory
+import org.odk.collect.maps.MapPoint
 import org.odk.collect.permissions.PermissionsChecker
 import javax.inject.Inject
