+++ geo/src/test/java/org/odk/collect/geo/GeoUtilsTest.java
@@ -9,7 +9,7 @@ import androidx.test.ext.junit.runners.AndroidJUnit4;
 
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.odk.collect.geo.maps.MapPoint;
+import org.odk.collect.maps.MapPoint;
 import org.odk.collect.testshared.LocationTestUtils;
 
 import java.util.ArrayList;
