+++ geo/src/test/java/org/odk/collect/geo/geopoint/GeoPointMapActivityTest.java
@@ -25,8 +25,8 @@ import org.odk.collect.geo.DaggerGeoDependencyComponent;
 import org.odk.collect.geo.GeoDependencyModule;
 import org.odk.collect.geo.R;
 import org.odk.collect.geo.ReferenceLayerSettingsNavigator;
-import org.odk.collect.geo.maps.MapFragmentFactory;
-import org.odk.collect.geo.maps.MapPoint;
+import org.odk.collect.maps.MapFragmentFactory;
+import org.odk.collect.maps.MapPoint;
 import org.odk.collect.geo.support.FakeMapFragment;
 import org.odk.collect.geo.support.RobolectricApplication;
 import org.robolectric.shadows.ShadowApplication;
