+++ geo/src/test/java/org/odk/collect/geo/selection/SelectionMapFragmentTest.kt
@@ -34,12 +34,12 @@ import org.odk.collect.geo.DaggerGeoDependencyComponent
 import org.odk.collect.geo.GeoDependencyModule
 import org.odk.collect.geo.R
 import org.odk.collect.geo.ReferenceLayerSettingsNavigator
-import org.odk.collect.geo.maps.MapFragment
-import org.odk.collect.geo.maps.MapFragmentFactory
-import org.odk.collect.geo.maps.MapPoint
 import org.odk.collect.geo.support.FakeMapFragment
 import org.odk.collect.geo.support.Fixtures
 import org.odk.collect.geo.support.RobolectricApplication
+import org.odk.collect.maps.MapFragment
+import org.odk.collect.maps.MapFragmentFactory
+import org.odk.collect.maps.MapPoint
 import org.odk.collect.permissions.PermissionsChecker
 
 @RunWith(AndroidJUnit4::class)
