+++ settings.gradle
@@ -24,3 +24,4 @@ include ':glide'
 include ':androidtest'
 include ':settings'
 include ':servicetest'
+include ':maps'
